tests/wpt/tests/css/css-color-hdr/parsing.html

<!DOCTYPE html>
<title>Parsing dynamic-range-limit with valid and invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-color-hdr/#the-dynamic-range-limit-property">
<meta name="assert" content="Parsing of dynamic-range-limit">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
  test_valid_value("dynamic-range-limit", "standard");
  test_valid_value("dynamic-range-limit", "no-limit");
  test_valid_value("dynamic-range-limit", "constrained");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 80%, standard 20%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 8%, standard 2%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 99%, standard 99%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 99%, standard 99%, constrained 10%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(constrained 20%, no-limit 80%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(dynamic-range-limit-mix(constrained 90%, no-limit 10%) 1%, no-limit 80%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 1%, dynamic-range-limit-mix(constrained 2%, no-limit 10%) 80%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(dynamic-range-limit-mix(no-limit 1%, standard 2%) 3%, dynamic-range-limit-mix(constrained 5%, no-limit 5%) 6%)");
  test_valid_value("dynamic-range-limit", "dynamic-range-limit-mix(standard calc(50% * sign(10em - 1px)), constrained 50%)");

  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 80%, standard 20%, )");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit, standard 20%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(constrained, no-limit, 80%)");
  test_invalid_value("dynamic-range-limit", "");
  test_invalid_value("dynamic-range-limit", "none");
  test_invalid_value("dynamic-range-limit", "default");
  test_invalid_value("dynamic-range-limit", "hdr");
  test_invalid_value("dynamic-range-limit", "sdr");
  test_invalid_value("dynamic-range-limit", "low");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 1%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 80% standard 20%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(low, no-limit, 10%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 101%, standard 1%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit -1%, standard 1%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(standard, no-limit, 0.1)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(no-limit 0%, standard 0%)");
  test_invalid_value("dynamic-range-limit", "dynamic-range-limit-mix(dynamic-range-limit-mix(no-limit 1%, standard 2%) 3%, dynamic-range-limit-mix(constrained 0%, no-limit 0%) 6%)");
</script>
