third_party/blink/web_tests/external/wpt/css/css-color-hdr/parsing-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['dynamic-range-limit'] = "dynamic-range-limit-mix(no-limit 1%)" should not set the property value
  assert_equals: expected "" but got "dynamic-range-limit-mix(no-limit 1%)"
Harness: the test ran to completion.
